Homework 3
