+++ src/main/java/Validation/ProductValidationService.java
@@ -1,5 +1,6 @@
 package Validation;
 import Domain.ProductVariables;
+import Product.ProductRepository;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -9,11 +10,15 @@ public class ProductValidationService {
     private Set<ProductValidationRule> validationRules = new HashSet<>();
     private Set<ProductValidationPercentRule> validationPercentRules = new HashSet<>();
     private Set<ProductValidationPriceRule> validationPriceRules = new HashSet<>();
+    private Set<ProductValidationUniqueRule> validationUniqueRules = new HashSet<>();
+
+
 
     public ProductValidationService() {
         validationRules.add(new ProductNameValidationRule());
         validationPercentRules.add(new ProductPercentValidationRule());
         validationPriceRules.add(new ProductPriceValidationRule());
+        validationUniqueRules.add(new ProductUniqueName());
     }
 
     public void validate(ProductVariables productVariables) {
@@ -26,5 +31,9 @@ public class ProductValidationService {
 
     public void validatePrice(ProductVariables productVariables) {
         validationPriceRules.forEach(s -> s.validatePrice(productVariables));
+    }
+
+    public void validateUniqueName(ProductVariables productVariables) {
+        validationUniqueRules.forEach(s -> s.validateUniqueName(productVariables));
     }
 }
